+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.43.1 (Unreleased)
 
 * Resource: `tencentcloud_instance` add new argument `encrypt` to support data disk with encrypt
+* Resource: `tencentcloud_elasticsearch` add new argument `encrypt` to support disk with encrypt
 * Resource: `tencentcloud_kubernetes_cluster` add new argument `cam_role_name` to support authorization with instances
 
 ## 1.43.0 (September 18, 2020)